TOBAGO-1976: Backport to 2.x: Using Servlet 3.0 for uploading files with <tc:file>

+++ tobago-example/tobago-example-demo/src/main/java/org/apache/myfaces/tobago/example/demo/UploadItem.java
@@ -27,10 +27,10 @@ public class UploadItem {
   private static final Logger LOG = LoggerFactory.getLogger(UploadItem.class);
 
   private String name;
-  private int size;
+  private long size;
   private String type;
 
-  public UploadItem(final String name, final int size, final String type) {
+  public UploadItem(final String name, final long size, final String type) {
     this.name = name;
     this.size = size;
     this.type = type;
@@ -44,7 +44,7 @@ public class UploadItem {
     this.name = name;
   }
 
-  public int getSize() {
+  public long getSize() {
     return size;
   }
